Mark repaired data tracking (in)conclusive rather than just tracking pending session ids

+++ test/unit/org/apache/cassandra/db/ReadCommandTest.java
@@ -613,6 +613,8 @@ public class ReadCommandTest
         testRepairedDataTracking(cfs, readCommand);
     }
 
+    // unrepaired data shrinks filter marks inconclusive - names
+
     private void testRepairedDataTracking(ColumnFamilyStore cfs, ReadCommand readCommand) throws IOException
     {
         cfs.truncateBlocking();
@@ -626,7 +628,7 @@ public class ReadCommandTest
 
         cfs.forceBlockingFlush();
 
-        new RowUpdateBuilder(cfs.metadata(), 0, ByteBufferUtil.bytes("key"))
+        new RowUpdateBuilder(cfs.metadata(), 1, ByteBufferUtil.bytes("key"))
                 .clustering("dd")
                 .add("a", ByteBufferUtil.bytes("abcd"))
                 .build()
@@ -645,51 +647,54 @@ public class ReadCommandTest
         // Capture all the digest versions as we mutate the table's repaired status. Each time
         // we make a change, we expect a different digest.
         Set<ByteBuffer> digests = new HashSet<>();
-        // first time round, nothing has been marked repaired so we expect digest to be an empty buffer
-        ByteBuffer digest = performReadAndVerifyRepairedInfo(readCommand, numPartitions, rowsPerPartition);
+        // first time round, nothing has been marked repaired so we expect digest to be an empty buffer and to be marked conclusive
+        ByteBuffer digest = performReadAndVerifyRepairedInfo(readCommand, numPartitions, rowsPerPartition, true);
         assertEquals(ByteBufferUtil.EMPTY_BYTE_BUFFER, digest);
         digests.add(digest);
 
-        // add a pending repair session to table1, digest should remain the same
+        // add a pending repair session to table1, digest should remain the same but now we expect it to be marked inconclusive
         UUID session1 = UUID.randomUUID();
         mutateRepaired(sstable1, ActiveRepairService.UNREPAIRED_SSTABLE, session1);
-        digests.add(performReadAndVerifyRepairedInfo(readCommand, numPartitions, rowsPerPartition, session1));
+        digests.add(performReadAndVerifyRepairedInfo(readCommand, numPartitions, rowsPerPartition, false));
         assertEquals(1, digests.size());
 
-        // add a different pending session to table2, digest should remain the same, but we expect to see both sessions now
+        // add a different pending session to table2, digest should remain the same and still consider it inconclusive
         UUID session2 = UUID.randomUUID();
         mutateRepaired(sstable2, ActiveRepairService.UNREPAIRED_SSTABLE, session2);
-        digests.add(performReadAndVerifyRepairedInfo(readCommand, numPartitions, rowsPerPartition, session1, session2));
+        digests.add(performReadAndVerifyRepairedInfo(readCommand, numPartitions, rowsPerPartition, false));
         assertEquals(1, digests.size());
 
         // mark one table repaired
         mutateRepaired(sstable1, 111, null);
-        // this time, digest should not be empty, and we should only see session2
-        digests.add(performReadAndVerifyRepairedInfo(readCommand, numPartitions, rowsPerPartition, session2));
+        // this time, digest should not be empty, session2 still means that the result is inconclusive
+        digests.add(performReadAndVerifyRepairedInfo(readCommand, numPartitions, rowsPerPartition, false));
         assertEquals(2, digests.size());
 
         // mark the second table repaired
         mutateRepaired(sstable2, 222, null);
-        // digest should be updated again and there should be no pending sessions reported
-        digests.add(performReadAndVerifyRepairedInfo(readCommand, numPartitions, rowsPerPartition));
+        // digest should be updated again and as there are no longer any pending sessions, it should be considered conclusive
+        digests.add(performReadAndVerifyRepairedInfo(readCommand, numPartitions, rowsPerPartition, true));
         assertEquals(3, digests.size());
 
-        // insert a partition tombstone into the memtable, then re-check the repaired info
-        // this is to ensure that optimisations which skip reading from tables where a newer
-        // partition tombstone is read are disabled when tracking repaired data. To not do so
-        // risks false positives caused by as-yet unrepaired partition deletes
-        new Mutation(PartitionUpdate.simpleBuilder(cfs.metadata(), ByteBufferUtil.bytes("key"))
-                                    .delete()
-                                    .build()).apply();
-        digest = performReadAndVerifyRepairedInfo(readCommand, 0, rowsPerPartition);
-        assertNotEquals(ByteBufferUtil.EMPTY_BYTE_BUFFER, digest);
-        assertTrue(digests.contains(digest));
+        // insert a partition tombstone into the memtable, then re-check the repaired info.
+        // This is to ensure that when the optimisations which skip reading from sstables
+        // when a newer partition tombstone has already been cause the digest to be marked
+        // as inconclusive.
+        // the exception to this case is for partition range reads, where we always read
+        // and generate digests for all sstables, so we only test this path for single partition reads
+        if (readCommand.isLimitedToOnePartition())
+        {
+            new Mutation(PartitionUpdate.simpleBuilder(cfs.metadata(), ByteBufferUtil.bytes("key"))
+                                        .delete()
+                                        .build()).apply();
+            digest = performReadAndVerifyRepairedInfo(readCommand, 0, rowsPerPartition, false);
+            assertEquals(ByteBufferUtil.EMPTY_BYTE_BUFFER, digest);
 
-        // now flush so we have an unrepaired table with the deletion and repeat the check
-        cfs.forceBlockingFlush();
-        digest = performReadAndVerifyRepairedInfo(readCommand, 0, rowsPerPartition);
-        assertNotEquals(ByteBufferUtil.EMPTY_BYTE_BUFFER, digest);
-        assertTrue(digests.contains(digest));
+            // now flush so we have an unrepaired table with the deletion and repeat the check
+            cfs.forceBlockingFlush();
+            digest = performReadAndVerifyRepairedInfo(readCommand, 0, rowsPerPartition, false);
+            assertEquals(ByteBufferUtil.EMPTY_BYTE_BUFFER, digest);
+        }
     }
 
     private void mutateRepaired(SSTableReader sstable, long repairedAt, UUID pendingSession) throws IOException
@@ -701,13 +706,12 @@ public class ReadCommandTest
     private ByteBuffer performReadAndVerifyRepairedInfo(ReadCommand command,
                                                         int expectedPartitions,
                                                         int expectedRowsPerPartition,
-                                                        UUID...expectedSessionIds)
+                                                        boolean expectConclusive)
     {
         // perform equivalent read command multiple times and assert that
         // the repaired data info is always consistent. Return the digest
         // so we can verify that it changes when the repaired status of
         // the queried tables does.
-        Set<UUID> expectedSessions = new HashSet<>(Arrays.asList(expectedSessionIds));
         Set<ByteBuffer> digests = new HashSet<>();
         for (int i = 0; i < 10; i++)
         {
@@ -722,8 +726,7 @@ public class ReadCommandTest
             ByteBuffer digest = info.getRepairedDataDigest();
             digests.add(digest);
             assertEquals(1, digests.size());
-
-            assertEquals(expectedSessions, info.getPendingRepairSessions());
+            assertEquals(expectConclusive, info.isConclusive());
         }
         return digests.iterator().next();
     }
